+++ frontend/Checkout/subscribers.js
@@ -1,10 +1,15 @@
 import goBackHistory from '@shopgate/pwa-common/actions/history/goBackHistory';
-import { isUserLoggedIn } from '@shopgate/pwa-common/selectors/user';
 import ParsedLink from '@shopgate/pwa-common/components/Router/helpers/parsed-link';
 import trackingCore from '@shopgate/tracking-core/core/Core';
 import { LEGACY_URL } from '@shopgate/pwa-common-commerce/checkout/constants';
 import { openedCheckoutLink$ } from '@shopgate/pwa-common-commerce/checkout/streams';
 import fetchCheckoutUrl from '@shopgate/pwa-common-commerce/checkout/actions/fetchCheckoutUrl';
+import { appDidStart$ } from '@shopgate/pwa-common/streams/app';
+import PipelineRequest from '@shopgate/pwa-core/classes/PipelineRequest';
+import { logger } from '@shopgate/pwa-core/helpers';
+import event from '@shopgate/pwa-core/classes/Event';
+
+const MARK_SHOPGATE_ORDER = 'bigcommerce.checkout.markOrderAsShopgate'
 
 /**
  * Checkout subscriptions.
@@ -15,10 +20,6 @@ export default function checkout(subscribe) {
    * Gets triggered when the user enters the checkout.
    */
   subscribe(openedCheckoutLink$, ({ dispatch, getState }) => {
-    // Check if user is logged in.
-    if (!isUserLoggedIn(getState())) {
-      return;
-    }
 
     dispatch(fetchCheckoutUrl())
       .then((url) => {
@@ -37,4 +38,19 @@ export default function checkout(subscribe) {
 
     dispatch(goBackHistory(1));
   });
-}
+
+  subscribe(appDidStart$, () => {
+    event.addCallback('checkoutSuccess', (data = {}) => {
+      if (typeof data.order === 'undefined') {
+        return;
+      }
+
+      const orderNumber = data.order.number;
+
+      new PipelineRequest(MARK_SHOPGATE_ORDER)
+        .setInput({ orderId: orderNumber })
+        .dispatch()
+        .catch(err => logger.error(err));
+    });
+  });
+};